+++ src/extension.ts
@@ -31,6 +31,8 @@ function setup(workspaceFolders: readonly vscode.WorkspaceFolder[]) {
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export async function activate(context: vscode.ExtensionContext) {
+  app.vscodeContext = context;
+
   try {
     initCommands(context);
   } catch (error) {